+++ src/utils.ts
@@ -268,13 +268,13 @@ export const setCursorForShape = (
   if (!canvas) {
     return;
   }
-  if (appState.elementType === "selection") {
+  if (appState.activeTool.type === "selection") {
     resetCursor(canvas);
-  } else if (appState.elementType === "eraser") {
+  } else if (appState.activeTool.type === "eraser") {
     setEraserCursor(canvas, appState.theme);
     // do nothing if image tool is selected which suggests there's
     // a image-preview set as the cursor
-  } else if (appState.elementType !== "image") {
+  } else if (appState.activeTool.type !== "image") {
     canvas.style.cursor = CURSOR_TYPE.CROSSHAIR;
   }
 };
@@ -611,4 +611,17 @@ export const updateObject = <T extends Record<string, any>>(
     ...obj,
     ...updates,
   };
+};
+
+export const isPrimitive = (val: any) => {
+  const type = typeof val;
+  return val == null || (type !== "object" && type !== "function");
+};
+
+export const getFrame = () => {
+  try {
+    return window.self === window.top ? "top" : "iframe";
+  } catch (error) {
+    return "iframe";
+  }
 };